Until API Implementation
Created Repository interface and and implemented all the data retrieval or adding functionality in it using Cars model.
Created DTOs and used on controller inorder to avoid exposure of our model with directly works with our database.
Created DTO based on need in API request.
Injected AutoMapper by installing thru nugget package.
Registered the dynamoDB in program.cs so that we can use anywhere in the project .
made separate file for it's configuration tho.